record_agent/requirements.txt
langchain-core==0.3.63
langchain==0.3.25
langchain-community==0.3.5
langchain-text-splitters==0.3.8
chromadb==0.5.23
dashscope==1.17.0
pypdf==4.3.1
python-docx==1.1.0
python-pptx==0.6.23
chardet==5.2.0
pydantic-settings==2.4.0
sqlalchemy==2.0.35
pymysql==1.1.0
python-dateutil==2.9.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
flask==3.0.0
fake-useragent==1.4.0
